Yet another fix for database migration code.
The old code was checking for an exception that would only be
thrown by some upgrade scripts. Now an explicit check is made for
the alembic_version table; if that doesn't exist, the schema is
created fresh. Otherwise, the upgrade is run. Unlike previously,
this commit ensures that only an upgrade *or* an initialise is
performed, not both.