Inactive user filter button is now tri-state.
The search bar for users now has a filter button with three states:
 - Show active users only
 - Show all users
 - Show inactive users only

This also required changing the way the user list was nested in
the organisation view. Now it's a directive that manages its own
scope. Because of this, the number of route resolution parameters
has been reduced significantly. Although this means the lists may
be initially empty, this is an improvement because it reduces the
number of trips to the server and also reduces the number of lines
of code.