Fixed bogus animation when clickin on navbar item.
A click handler had been added to the entire navbar to cause it to
collapse when an item was clicked on. However it was also being
triggered when in desktop (large window) mode, which caused the
collapse animation to play even when the nav bar was never
collapsed. Now an additional CSS selector has been added so that
it doesn't animate unless it's on a small screen.