Added colour swatches next to hierarchy header.
One of the requirements for Aquamark is to have colour to identify
functions, processes, sub-processes and measures. This commit adds
colour swatches next to the headers. Each level has its own colour,
And also includes a small bit of text for visual identification
(e.g. "F" for functions, "Sp" for sub-processes).