Can now reorder functions of a survey.
The child list table has been refactored out into a directive for
reuse by other pages.

The Editor needed a small change for it to be able to call `reorder`
for arrays (e.g. the result of a query): now it takes an additional
parameter, which is the resource class. If the user tries to save
an array, the resource class is used to call `reorder` instead of
the instance method. Code that doesn't try to save arrays doesn't
need to pass in the resource class.