Using composite primary keys for questions.
Now that the question tables are not versioned, we need to use
composite primary keys in order to track the lineage of questions.
For example, two Functions may share the same ID but have different
survey IDs - in that case, they are logically the same function
but may have different content.

This also required the foreign keys to be modified to use the new
composite primary keys.