Fixed warnings in model relationships.
Added detailed primaryjoin parameters for relationships that join
on composite keys. This is required because such objects often have
multiple relationships that might both write to the same column.

http://docs.sqlalchemy.org/en/rel_1_0/orm/join_conditions.html#overlapping-foreign-keys

Now the mappings are constructed without warnings. It still needs
to be tested, though.

Extended the survey structure test case a little, but it doesn't
pass yet.