Removed old utils to_dict, normalise, simplify.
These have been replaced by ToSon, which is more powerful.

Also removed the *ListHander classes, because it's difficult to
define mappings like that: a URL such as /survey.json would need
to handle GET requests for a list of surveys, but a POST request
for a new survey. Calling the handler SurveyListHandler would
not make sense, so now all survey-related methods are in the one
handler, as it used to be for the other handlers.