Can link measures to multiple qnodes in GUI
Previously, there was only an Add button in the list of measures
when viewing a question node. This would let you create a new
measure. Now there is a Link button, which lets you search for
existing measures. They're presented in a table. Clicking on a
table row links that measure to the current qnode.