Added survey duplication function.
The method for creating a new survey can now take a duplicateId
parameter:

    POST /survey.json?duplicateId=:surveyId

The title and description will be taken from the request body,
just like creating a brand new survey. But then the structure
(hierarchy, qnodes, measures) will be copied from the referenced
survey.

Using ORM object traversal, because it ensures the rows are created
in the right order.