Can now open and close surveys for responding and editing.
Responding is not implemented yet, but the mechanism to open a
survey is. The controls are exposed as buttons on the Survey
screen.

When a survey is not editable, buttons such as Edit, Delete,
Unlink, Reorder and Add are not visible in the user interface -
and PUT methods will fail.