Added Aquamark response type metadata.
Added a JSON file that describes the response types as defined in
the Aquamark Excel file. The response types give both the scores
and business logic (e.g. predicates) for each option.

Added a response_types field to Survey. For Aquamark surveys, this
will just be the contents of the JSON file - but other surveys may
have different types.