Made qnode.hierarchy_id non-null for easy lookups.
Since responses don't link directly to rnodes, we need to traverse
the qnode tree to find the parent. But going from a measure to a
qnode is tricky, because there may be several of them. The solution
is to set the hierarchy ID on every qnode, so that the parent can
be looked up based on the assessment's hierarchy foreign key.
Previously, the hierarchy ID was only set on root qnodes.

To do: creating a new qnode as the child of a hierarchy fails for
some reason.