Categories can be marked as not relevant.
Functions, processes and sub-processes can now be marked as not
relevant. Previously, only measures could me maked so. This means
it is now possible for a respondent to opt out of entire sections
of the survey. Their score for those sections will be zero.

This is stored as metadata on the rnode; nothing deeper in the
structure is changed. This means it's possible to toggle the NA
state of a category without losing data that has already been
entered.

The NA state is shown in the list of child categories, so it's
easy to see which categories have been marked as NA. A statistics
summary has also been added to the top of the category screen
that shows the total score, number of measures that have been
completed, and the number of measures that are marked as NA.