Assessment duplication now maintains NA status of categories.
Added code to duplicate rnodes as well as responses. The scores
are recalculated after that, but the NA status of each category is
kept. Rnodes will be dropped if there is no matching qnode in the
target hierarchy.

Setting relations instead of IDs, since this seems to be more in
line with how SQLAlchemy wants to work.

Added unique constraints to rnode and response tables to prevent
multiple rnodes from being created for each qnode. This could
happen previously; that bug is now fixed.