Moved Submission selection screen inside Survey.
Previously, the list of submissions (assessments) was given on the
Program screen. Now it's only shown after selecting a survey. This
should make it clear that submissions are given in relation to a
particular survey.

Also, the bar at the top of the screen that shows which submission
you are viewing now has a button that allows you to quickly switch
to another submission without losing your place in the tree.

Still to do: filter the list of submissions to only show those for
the selected survey.