Displaying hierarchy labels next to sub headings
Previously, the short labels associated with each level of the
hierarchy were displayed next to the headings in the blue headings
bar. Now they are also displayed in the sub-sections (where the
user can choose which category to open next). This gives the user
a visual cue about what label to expect to see after they click on
an item.

Currently only the label text is displayed; the colours are not.