Using purchased_survey table instead of survey.open_date.
Survey ownership is denoted by a many-to-many relationship given
by the purchased_survey table. This gives an organisation a
collection of surveys that it may respond to. This information is
now used to determine whether a new submission can be created for
an organisation.

Currently, once a submission has been created it can be filled out
even if survey ownership is later revoked.