When linking to a measure in the same hierarchy it is moved
Previously, it was possible to have two copies of a measure in the
same hierarchy. Now, when a measure is linked to a qnode, any old
links within the same hierarchy are removed.

Also added a mechanism for the server to send messages to the
client as response metadata. In this instance, the user is notified
of the type of operation that was actually performed (move or just
link).