Added a health check service to fix AWS load balancer.
Load balancers on AWS make periodic requests to the backend servers
to make sure they are still running. The default is /index.html,
but our /index.html can not be viewed without authentication.

This commit adds a new service at /ping that checks that the
database is accessible and then returns 200 OK. The user does not
need to be logged in.