Fix for XSRF cookie expiring too early
The XSRF token is set in the login handler. Because that's before
the _current_user is set, the cookie is set to expire when the
browser session expires. This mismatch with the user login expiry
causes much confusion, because the user may be logged in but they
can't perform POST requests.