Adhoc: Using read-only user for arbitrary queries
A user with restricted permissions is created when the application
initialises the database. The password for that user is randomly
generated, and stored in the system config table. That information
is used to create a new connection string to use for the restricted
user. A read-only session factory is created using the restricted
user's credentials.

The read-only session is used to execute ad-hoc queries, so it's
not possible for a user to accidentally delete data. Some tables
and columns are also protected against being viewed.