Removed `parent` parameter from response URL
When viewing a measure, a `parent` parameter is required, because
the measure may have multiple parents. This is not the case when
viewing the measure as part of an assessment, because the measure
can only have one parent in any given hierarchy. Removing the
`parent` parameter from the URL allows a response to be properly
viewed when clicking on it in the timeline.