Organisation metadata split off into separate classes/tables
Added OrgMeta and OrgRegion tables. org_meta stores information
about water utility-related stuff, such as the number of customers.
org_region stores geographical information, such as country and
city.

Also added a web service that communicates with OSM's Nominatim
service to search for cities and retrieve hierarchical information.
The results are cached to reduce the load on the Nominatim server.