Fixed user creation, and test cases.
User POST method was missing call to flush, which is required to
generate ID before creating an entry in the action log (and also
required to perform the following call to get).

Also fixed a couple of test cases that weren't expecting the
deleted field in the JSON returned by the web services.