Added bulk appoval buttons to rnode page
There is now a button on submission category screen that exposes
buttons for running bulk approvals. If the user clicks one of the
buttons, all responses in that category will have their approval
state set to the user's choice. Options are:

 - Promote, demote, or promote and demote responses to the chosen
   state
 - Create (or not) missing responses and mark them as Not Relevant

Also added a new service and directive for highlighting an area of
the screen that needs attention, like a diagolue box but in-page.