Small refactor of some test cases, and made score sums unambiguous
Instead of using 100, 200, 300 for the measure weights in the tests,
numbers have been chosen such that all combinations give unique
sums. This makes it easier to verify the results of a test.

The refactor was to split up the survey modification test, again
so that it's easier to know what state the data should be in.