Fixed chart/statistics approval filtering
The approval order was wrong.

Also, the approval buttons on the chart screen now use the
approval-buttons widget. This widget has been given a new mode
which draws greater-or-equal states as active. Previously, the
chart screen didn't accurately show which states were active,
giving the impression that only submissions with the selected state
were included in the result.