Removed delay for submission category (rnode) state
The importance, urgency and NA fields of an rnode used to have a
delay of 1.5s: after clicking on a button, you had to wait 1.5s
before the changes were applied. This causes confusion when moving
between screens, because the new screen may still be displaying
the old information.

Now there's no delay, which means more requests but less confusion.