Fixed _xsrf cookie to expire at the same time as user cookie
Previously, the _xsrf cookie was being set when the login page
loaded, but the other cookies were being set when the user actually
logged in. This meant the _xsrf cookie could expire an arbitrary
amount of time before the user's session, which would make it
impossible to properly log in.

Now, the _xsrf cookie is set at the same time as the user cookie,
and they are given the exact same expiry time instead of a time
offset.