Fixed next/previous buttons for multi-surver programs
When browsing the first level of categories in a survey, clicking
next might actually take you to a category in a different survey.
This was because the database query was underconstrained: it
assumed that, since qnode IDs are not shared between surveys, if
two qnodes had the same parent ID they must belong to the same
survey. That's true for deeper categories, but for the first level
the parent ID is null, which can indeed be shared across surveys.