Counting number of draft responses in a survey
The survey progress meters now also show the number of draft
responses in addition to the final, reviewed and approved
responses.

A new column was added to the table, the web services were updated,
an upgrade script was created, and the user interface was updated
to include the extra bar. The upgrade is quite slow: about 4s per
assessment. Progress is printed to the console.