Added HTML sanitation on the server
Description fields are now sanitised on the server to remove
malicious HTML code. This is done using the 3rd-party bleach
module.

This involved a refactor of utils.ToSon: now, the initialiser
takes a list of expressions, each of which can be prefixed with
a special control character:

 - '' (nothing) - add expression to include list
 - '<' - add expression to include list, and sanitise HTML
 - '!' - add expression to exclude list

Previously, separate lists had to be passed in, and it was not
possible to sanitise a field.

Also, utils.updater has had a `sanitise` option added to it.