Improved markdown conversion, and added conversion log
Added a conversion log so that changes to a survey may be checked.
The log files are written to /tmp/aq/*.log, and can be compared
using e.g.

    colordiff -U3 /tmp/aq/{orig,new}_desc.log -b -F '^-- /#' | less -R

Also improved the conversion of lists to be more tolerant of
inline lists.

IMPORTANT:
There are however some cases that it doesn't handle, for example,
some sublists are preceded by a paragraph that should be indented
but which is not. These will need to be manually fixed. So run the
upgrade, then manually edit the qnodes and measures in the UI.