Using modular classes for response score calculation in GUI
The GUI code has been updated to match the design of the
response score calculation on the server. They are now very similar
indeed, the intention being that they should always be kept in
sync with one another.

The response GUI has been updated to use the new classes.

This commit adds support for score calculation using numerical
response parts in the GUI, but the actual controls for displaying
such response parts has not been implemented yet.