Moved response form into a column on the right
It was inconvenient to have to scroll to the bottom of the page
every time you changed to a different measure. Now the response
form and measure description are displayed side-by-side, on screens
that are big enough.

Also stopped using transform to scale active buttons, because that
causes the text to become blurry. Instead, a box shadow is used to
make the button look bigger.