Using qnode_measures instead of parents on measures
A Measure may have many parents. This is stored as a many-to-many
mapping via the qnode_measure table. QnodeMeasure objects were
historically used to represent this relationship, and a little
metadata for the sequence number of the measure within the parent
qnode's collection. Association proxies were used to give two-way
access via the qnode.measures and measure.parents properies.

Now, those properties have been removed, and explicit access is
required. Use the qnode.qnode_measures and measure.qnode_measures
properties. This is because QnodeMeasure is not really (and has
never been) just a many-to-many mapping: it's the incarnation of
a measure within a survey, and without it, a survey belongs only
to a program. This change is required now because it doesn't make
sense to bind responses or variables to a measure object; instead
they must be bound to a qnode_measure.