Added error handling for survey and submission dependencies
Added an `error` field to several survey structure and submission
tables, which is used to store errors for display to the user.
It's filled in when there is a problem updating the survey score
or calculating the dependency graph.

Added unit tests to test error functionality.

Fixed some bugs in submission score calculation code.