Added response type viewing screen
So that you can click on the response types in the activity stream.
Currently it is a very simple page that just shows how many
measures use the response type. There is a link to the list of
measures. It's not possible yet to edit the response type from
the new screen; that must be done on the measure screen.