app_base_url is now a config parameter available in admin user interface
This parameter was previously written into the config files, by including it in the DB config options it can be altered through the user interface and allows multiple instances to share the same config files. The parameter is used in the notification.yaml file to formulate the notification unsubscribe link.