Allowed consultants to generate temporal reports for draft submissions
The temporal report was previously restricted such that draft and
final submissions were excluded, and only admin users could request
a report that included all submissions. Now consultants can do
that too.

The reason is that consultants need to review all the data even
before it has been made final, and it's very difficult to do without
this feature.