Partial conversion of many HTTP handlers to new AuthZ scheme
Added a UserSession class, which can be obtained by calling
BaseHandler.get_user_session(). It replaces a bunch of other
attributes of BaseHandler, such as `current_user`. The UserSession
is tied to a database session, so ORM functions work on the
objects it contains. This enables the implementation of
AuthZ-supporting functions such as `UserSession.purchased_survey`.

More work is required; see the changes to `BaseHandler` and make
sure other classes are using the new interface.