Using roles in authz.yaml as primary definition
Previously, roles.json contained names and descriptions of the
roles available in the application. Now this is contained only
in authz.yaml, so that the documentation is all in one place.
Also, the role hierarchy is defined using AuthZ rules instead of
arrays hard-coded all over the place.

authz.Policy.check now takes a variable number of rule names. By
default, it returns true if any of the rules succeed.