Added survey group restrictions for timeline
Actions now have survey groups associated with them, so the
timeline can be filtered based on the current user's groups.
This unfortunately means that the groups of each action are
static: if a program changes groups, old actions relating
to it will still be visible via the old groups. However, the
program will still not be accessible via the old actions.

Added a choice of survey groups to the action post screen.