Added a more explicit help message for consultants viewing surveys
On the submission choice screen, a messgae is displayed telling
the user that they need to purchase the survey before creating
a submission if they haven't done so yet. But this was confusing
for consultants, who can act on behalf of other organisations.

Now there's another message informing them that they can switch
organisations.