Bug fixes for box and whisker charts and orphaned categories
Box and whisker charts were plotting off the page if measure scores
were not restricted to be between 0 and 1. Fixed by setting chart
maximum y-value as the larger of the totalWeight or the maximum
submission score.

Orphaned categories were preventing submissions from being marked
as final. The server side test causing this to fail was not filtering
deleted categories from it's check. This has been fixed.