Bug fix: Gave analyst read permissions for response_type table
This was causing issues in custom query to compute "median value
for the scores for each of the four response types for each of the
7 functions".

The query now only fails due to execution time constraints enforced
in the app configuration. I think these may be hard-coded in the
CustomQuerySettings service.